CGI works, refactor in progress
