+++ .gitignore
@@ -1,3 +1,4 @@
 .vscode
 *.o
 webserv
+.idea
