+++ src/utils/CGIUtils.hpp
@@ -19,3 +19,4 @@ void setResHeader(std::string response, http::Response &res);
 void setResBody(std::string response, http::Response &res);
 std::string getInterpreter(const std::string& path, const std::map<std::string, std::string>& extMap);
 bool isCgi(const std::string &path, const std::map<std::string, std::string>& extMap);
+std::vector<std::string> split(const std::string& str, char delimiter);
